Изменена обработка координат в системе: поддержка чисел с плавающей запятой (float). в FileParser.cpp и FileParser.h
Описание:

Изменение типа данных для координат
Все структуры, работающие с координатами (в частности, Point), были обновлены для работы с числами с плавающей запятой (float) вместо целых чисел (int). Это изменение позволяет обрабатывать координаты точек и вершин, которые могут быть нецелыми числами (например, 3.14).

Обновление парсинга данных

В методах парсинга строк, таких как parsePointLine, был заменён механизм преобразования строк в числа с использованием std::stof, чтобы поддерживать числа с плавающей запятой.

Ранее использовавшийся std::stol для преобразования в long был заменён на std::stof, что позволяет корректно обрабатывать числа типа float.

Корректировка диапазонов для float
Все проверки, которые касаются диапазонов значений для координат, были обновлены для использования диапазона float (например, проверка на допустимость значений от -999.0 до 999.0).

Обновление логики проверки и обработки ошибок

В функциях, таких как checkOutOfRangeCoordinates, теперь проверяются значения типа float для корректного диапазона.

Сообщения об ошибках были обновлены с учётом работы с числами с плавающей запятой (например, сообщения теперь сообщают о нарушении диапазона для координат типа float).

Функциональные изменения

Изменения затронули методы работы с файлами (например, readFromFile), где теперь поддерживаются координаты типа float.

Добавлены дополнительные проверки для корректного распознавания и обработки данных с плавающей запятой при чтении координат из файла.

Причина изменений:
Эти изменения были сделаны для того, чтобы расширить поддержку работы с точными координатами, которые могут быть нецелыми числами (например, с плавающей запятой), улучшив точность работы программы и её гибкость при обработке данных.

+++ FileParser.h
@@ -39,10 +39,10 @@ private:
     bool checkNonIntegerCoordinates(const std::string& line, Error& err, int lineNumber);
 
     /// Парсинг строки "x;y" → два int (возвращает false, если не удалось)
-    bool parsePointLine(const std::string& line, int& x, int& y);
+    bool parsePointLine(const std::string& line, float& x, float& y);
 
     /// Проверка, что координаты (x, y) лежат в диапазоне [-999, 999]
     /// isVertex = true → проверяем вершину, иначе — проверяемая точка
-    bool checkOutOfRangeCoordinates(int x, int y, Error& err, int lineNumber, bool isVertex);
+    bool checkOutOfRangeCoordinates(float x, float y, Error& err, int lineNumber, bool isVertex);
 };
